move init, add app name, extend future struct
