+++ public/less/main.less
@@ -2,3 +2,6 @@
   margin-top: 30px;
 }
 
+.navbar-inverse .logo {
+  width: 32px;
+}
