+++ server/routes/config.js
@@ -1,0 +1,3 @@
+module.exports = function (request, reply) {
+  reply(require('../lib/config'));
+};
