+++ server/routes/items.js
@@ -1,0 +1,4 @@
+const config = require('../lib/config');
+module.exports = function (request, reply) {
+  reply([]);
+};
